front : refactor, fix bug (delete book not working in give-up list)
