+++ frontend/src/components/book/HorizontalBookView.js
@@ -1,43 +1,12 @@
 import React from 'react'
-import { Button } from 'react-bootstrap'
-import toast from 'react-hot-toast'
 // Book Info
 import BookInfoIcon from './book-info/BookInfoIcon'
 // Images
 import defaultBookCover from '../../resources/images/common/book.png'
 import { CATEGORY_INFO, FORM_INFO, LANGUAGE_INFO } from './book-info/bookInfoEnum'
 import PageProgressBar from '../common/PageProgressBar'
-import { giveUpBook } from '../../functions/book'
 
-const HorizontalBookView = ({
-	book,
-	firstButton = (
-		<a href={`/reading/${book.bookId}`} className='btn btn-primary w-100'>
-			이어서 읽기
-		</a>
-	),
-	secondButton = (
-		<Button
-			variant='danger'
-			className='w-100'
-			onClick={() => {
-				const confirm = window.confirm('책을 포기할까요?')
-
-				if (confirm) {
-					giveUpBook(book.bookId).then((success) => {
-						if (success) {
-							toast.success('책을 포기했어요. 마음이 언제든지 다시 시작하실 수 있어요!')
-						} else {
-							toast.error('오류가 났어요 다시 시도해 주세요')
-						}
-					})
-				}
-			}}>
-			이 책 포기하기
-		</Button>
-	),
-	link = '',
-}) => {
+const HorizontalBookView = ({ book, firstButton = <></>, secondButton = <></>, link = '' }) => {
 	const bookInfoStyle = 'col-4 col-md-4 mb-2'
 
 	return (
